Project Update
Introducing the CustomerSearchView and the ProductSearchView (CustomerSearchApp and ProductSearchApp in test.views. for preview implementations). The -SearchViews both provide search support. The table views support direct cell editing.
TableView is extended as SaveTableView, which keeps track of all changes that are directly applied. The TableView subclass stores all affected object and keeps track of the applied changes which are stored as a stack. This data structure allows us to implement the undo system.